requirements.txt
wheel
aiohttp==3.7.4.post0
colorthief==0.2.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus
discord.py @ git+https://github.com/iDevision/enhanced-discord.py#egg=discord.py[voice]
emojis==0.6.0
googletrans==3.1.0a0
humanize==3.12.0
motor==2.5.1
parsedatetime==2.6
pymongo[srv]
Pillow==8.3.2
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0